chore(fpnv): Refactor pnv error logic
