+++ src/phone-number-verification/token-verifier.ts
@@ -23,7 +23,8 @@ import {
   DecodedToken, decodeJwt, JwtError, JwtErrorCode,
   PublicKeySignatureVerifier, ALGORITHM_ES256, SignatureVerifier,
 } from '../utils/jwt';
-import { FirebasePhoneNumberTokenInfo, FPNV_ERROR_CODE_MAPPING } from './phone-number-verification-api-client-internal';
+import { FirebasePhoneNumberTokenInfo } from './phone-number-verification-api-client-internal';
+import { FPNV_ERROR_CODE_MAPPING } from './error';
 
 export class PhoneNumberTokenVerifier {
   private readonly shortNameArticle: string;
